fix(dal): Rename mongo -> mongoose
