+++ libs/dal/src/types/index.ts
@@ -1,4 +1,4 @@
 export * from './enforce';
 export * from './helpers';
 export * from './results';
-export * from './mongo';
+export * from './mongoose';
